backend/requirements.txt
slack_bolt
gspread
oauth2client
pandas
mysql-connector-python
serverless-wsgi